Added function in board class to determine board size and calculate total cards and total matches. Called the function in the Memory class.
In Memory class added board.totalCards, player1.name, player2.name to parameters to pass those values to the getMatchedStatus() function. So that way the messages use the player's names and uses the total cards to calculate matches left. I made changes in the GetMatchedStatus class to reflect these changes as well.

I had to comment out a section in the Game class because it was causing an error. I think the changes I made in the Board class covered all that information, but I didn't want to delete any of Janis's code without talking to her first.

Also made a few changes in the Player class so make the messages display on a new line so readability while running the program was easier.